fastapi-sqlapi/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pyodbc>=4.0.32
python-dotenv>=0.19.0
aiohttp>=3.8.1